Factor out getActivitiesHandler
