+++ backend/api/handlers.go
@@ -347,22 +347,3 @@ func getIssuesHandler(c *fiber.Ctx) error {
 
 	return proxy.Do(c, redmineURL)
 }
-
-// getActivitiesHandler godoc
-// @Summary	Proxy for the "/enumerations/time_entry_activities.json" Redmine endpoint
-// @Accept	json
-// @Produce	json
-// @Failure	401	{string}	error "Unauthorized"
-// @Failure	500	{string}	error "Internal Server Error"
-// @Router /api/activities [get]
-func getActivitiesHandler(c *fiber.Ctx) error {
-	if ok, err := prepareRedmineRequest(c); !ok {
-		return err
-	}
-
-	redmineURL := fmt.Sprintf("%s/enumerations/time_entry_activities.json",
-		config.Config.Redmine.URL)
-
-	// Proxy the request to Redmine
-	return proxy.Do(c, redmineURL)
-}
